refactor: move main state provider to its own file
